Rename: Loader컴포넌트 src/component/common으로 이동
