+++ webapp/src/pages/Callback/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Loader from 'components/Loader';
+import Loader from 'components/Common/Loader';
 
 export default function Callback() {
   return <Loader />;
